requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
fastapi==0.115.14
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.0.1
python-multipart==0.0.22
python-telegram-bot==22.6
PyYAML==6.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.46.2
twilio==9.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
